requirements.txt
assertpy
cerberus
pillow
pytest
requests
streamdeck
